% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastGDCData}
\alias{breastGDCData}
\title{Count data from GDC Data Portal}
\format{
Objects of class \verb{data frame}.
}
\source{
GDC Data Portal, \url{https://portal.gdc.cancer.gov/}
}
\usage{
data(breastGDCData)
}
\description{
Count data composed by 1095 breast cancer patients. The data are filtered and nromalized by using \code{TCGAbiolinks} package.
The survival information include \code{time} (years) and \code{status} for each patients.
}
\examples{
data(breastGDCData)
}
\references{
Colaprico A, Silva TC, Olsen C, Garofano L, Cava C, Garolini D, et al. (2015).
TCGAbiolinks: An R/Bioconductor package for integrative analysis of TCGA data.\cr
\emph{Nucleic Acids Research} (\href{http://doi.org/10.1093/nar/gkv1507}{DOI}). \cr \cr
Mounir M, Lucchetta M, Silva C T, Olsen C, Bontempi G, Chen X, et al.(2019).
New functionalities in the TCGAbiolinks package for the study and integration of cancer data from GDC and GTEx. \cr
\emph{PLoS computational biology, 15(3), e1006701.} \cr \cr
}
\keyword{datasets}
