% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastGEOData}
\alias{breastGEOData}
\title{Microarray data from GEO database}
\format{
Objects of class \verb{data frame}.
}
\source{
GEO Database, \url{https://www.ncbi.nlm.nih.gov/geo/}
}
\usage{
data(breastGEOData)
}
\description{
Affymetrix HG-U133A microarrays of two different breast cancer datasets are downloaded from GEO database:
GSE2034 as training set and GSE2990 as testing set. The data are noromalized and annotated.
The \verb{training data} is composed by gene expression profiles from total RNA of frozen tumour samples from 286 lymph-node-negative breast cancer patients.
The \verb{testing data} is composed by gene expression profiles of of 189 invasive breast carcinomas.
The survival information include breast cancer relapse free survival (RFS): \code{time} (months) and \code{status} for each patients in the two datasets. `
}
\examples{
data(breastGEOData)
}
\references{
Wang Y, Klijn JG, Zhang Y, Sieuwerts AM et al.
Gene-expression profiles to predict distant metastasis of lymph-node-negative primary breast cancer.
#' \emph{Lancet 2005 Feb 19-25;365(9460):671-9} (\href{https://doi.org/10.1016/S0140-6736(05)17947-1}{DOI}). \cr \cr
Sotiriou C, Wirapati P, Loi S, Harris A et al.
Gene expression profiling in breast cancer: understanding the molecular basis of histologic grade to improve prognosis
#' \emph{J Natl Cancer Inst 2006 Feb 15;98(4):262-72} (\href{https://doi.org/10.1093/jnci/djj052}{DOI}). \cr \cr
}
\keyword{datasets}
