% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosmonetTraining.R
\name{CosmonetTraining}
\alias{CosmonetTraining}
\title{Fit network-regularized Cox regression models on training set}
\usage{
CosmonetTraining(
  k,
  x,
  y,
  screenVars,
  family = "Cox",
  penalty = "Net",
  Omega,
  alpha = 0.5,
  nfolds = 5
)
}
\arguments{
\item{k}{times to loop through cross validation.}

\item{x}{input training matrix \eqn{nxp}. Each row is an observation vector.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{screenVars}{screened variables obtained from BMD- or DAD-, or BMD+DAD-screening.}

\item{family}{Cox proportional hazards regression model.}

\item{penalty}{penalty type. Can choose "Net" where \code{Omega} matrix is requested.
For \code{penalty = Net}, the penalty is defined as
\eqn{\lambda*{\alpha*||\beta||_1+(1-\lapha)/2*(\beta^{T}L\beta)}},
where \eqn{L} is a Laplacian matrix calculated from \code{Omega}.}

\item{Omega}{adjacency matrix with zero diagonal and non-negative off-diagonal used to calculate Laplacian matrix.}

\item{alpha}{ratio between \code{L_1} and Laplacian for \code{Net}. Default is \code{alpha = 0.5}.}

\item{nfolds}{number of cross-validation performed for tuning optimal parameters over runs. Default is \code{nfolds = 5}.}
}
\value{
The following objects are returned:
\item{beta}{a sparse Matrix of coefficients, stored in class\code{dgCMatrix}.}
\item{opt.lambda}{value of \eqn{lambda} based on minimum \eqn{cvm} over runs.}
\item{df}{data frame of sample, prognostic index, time, status and group risk for each quantile \eqn{q_{gamma}}, with \eqn{\gamma=0.20, ..., 0.80}.}
\item{summary}{summary information about number of patients at risk, cutoff and p.value for each quantile.}
\item{p.value}{\eqn{p}-value resulting from the log-rank test (the significance level is \eqn{p}-value < 0.05).}
\item{plots}{survival curves and distribution plot of prognostic index \eqn{PI^{T}}.}
}
\description{
This function fits penalized Cox regression methods in order to incorporate gene regulatory relationships and to select signature genes using the training set \eqn{T}.
}
