% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAdjMatrix.R
\name{CreateAdjMatrix}
\alias{CreateAdjMatrix}
\title{Create 0/1 matrix by using the tissue-specific genes from DiseaseGenesSelection() function}
\usage{
CreateAdjMatrix(
  DiseaseGenes,
  tissueSpecificEdge,
  FLth = 0.5,
  output = TRUE,
  message = TRUE
)
}
\arguments{
\item{DiseaseGenes}{set of genes from the DiseaseGenesSelection() function or as input from the user - as entrez gene ids}

\item{tissueSpecificEdge}{file downloaded from https://hb.flatironinstitute.org/download. Top Edges: the network is filtered to only include edges with evidence supporting a tissue-specific functional interaction
[entrez gene id 1][entrez gene id 2][posterior prob.]}

\item{FLth}{FL adaptive threshold, default is 0.5 - edges greater than 0.5 are kept in the final graph}

\item{output}{if TRUE the Omega matrix is saved as .txt file in the current directory}

\item{message}{if TRUE returns a message showing the number of edges in the graph with the selected FLth threshold}
}
\value{
The following object isreturned:
\item{0/1}{adjcent matrix} of dimention equal to the number of genes selected as tissue-specific genes
}
\description{
This function creates the 0/1 network using the files available from https://hb.flatironinstitute.org/download
Since this is a cancer-specific study, the tissue-specific network has been downloaded.
The example below is done with the mammary epithelium top-edges Network file.
}
