% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetworkMatrix.R
\name{CreateNetworkMatrix}
\alias{CreateNetworkMatrix}
\title{Create network matrix by using the cancer-specific genes}
\usage{
CreateNetworkMatrix(
  x,
  repositoryDisease,
  diseaseID = NULL,
  topGenes = NULL,
  tissueSpecificEdge,
  output = TRUE,
  message = TRUE
)
}
\arguments{
\item{x}{input matrix \eqn{nxp}. Each row is an observation vector.}

\item{repositoryDisease}{the repository disease where the prediction score associated to each gene is a posterior probability. The higher the probability the stronger the functional relation between the gene and the cancer.}

\item{diseaseID}{the disease ID from disease ontology (http://disease-ontology.org).}

\item{topGenes}{the number of the top functional genes associated with the disease.}

\item{tissueSpecificEdge}{file downloaded from https://hb.flatironinstitute.org/download. Top Edges: the network is filtered to only include edges with evidence supporting a tissue-specific functional interaction.
[entrez gene id 1][entrez gene id 2][posterior prob.]}

\item{output}{if TRUE the Omega matrix is saved as .txt file in the current directory.}

\item{message}{if TRUE returns a message showing the number of edges in the graph.}
}
\value{
The following objects are returned:
\item{DiseaseGenes}{top cancer-disease genes and relative posterior probabilities.}
\item{Omega}{network matrix with zero diagonal and non-negative off-diagonal \eqn{pxp}.}
}
\description{
This function creates a network matrix with zero diagonal and non-negative off-diagonal used to calculate Laplacian matrix and of dimention \eqn{pxp}.
Since this is a cancer-specific study, the tissue-specific network has been downloaded.
The example below is done with the top-edges Network file using the files available at https://hb.flatironinstitute.org/download.
}
