% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseaseGeneSelection.R
\name{DiseaseGenesSelection}
\alias{DiseaseGenesSelection}
\title{Selection of cancer-related genes}
\usage{
DiseaseGenesSelection(repositoryDisease, diseaseID, topGenes = 500)
}
\arguments{
\item{repositoryDisease}{the repository disease where the prediction score associated to each gene is a posterior probability. The higher the probability the stronger the functional relation between the gene and the cancer.}

\item{diseaseID}{the disease ID from disease ontology (http://disease-ontology.org)}

\item{topGenes}{the number of the top functional genes associated with the disease - by defult is 500}
}
\value{
The following object is returned:
a data frame of the disease genes which is saved as .txt file in the current directory.
}
\description{
This function uses the file disease.partaa.gz to select the disease under investigation - available from https://humanbase-dev.s3-us-west-2.amazonaws.com/humanbase-data/data/predictions/disease.partaa.gz.
It returns a list of tissue-specific genes, the prediction scores as posterior probabilities and the disease IDs from disease ontology (http://disease-ontology.org).
and the gene IDs are Entrez.
}
