% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetwork.R
\name{GenerateNetwork}
\alias{GenerateNetwork}
\title{Generate a gene-network to perform a KEGG pathway analysis on the genes selected by the Cox-network based model}
\usage{
GenerateNetwork(
  ListOfGenes = NULL,
  header = TRUE,
  diseaseID = NULL,
  nodesCols = c("#EFCFD4", "#D9E3FC", "#FFFFFF")
)
}
\arguments{
\item{ListOfGenes}{a two-column data frame composed by gene names and beta coefficients}

\item{diseaseID}{the disease ID identified by Disease Ontology (DO) (https://disease-ontology.org)}

\item{ListHeader}{true if the genes' list has a header, false otherwise}

\item{nodesCol}{the colours of the nodes. Default values are orange, greeen and purple for mapped-up, mapped down and not mapped genes, respectively.}
}
\value{
Create KEGG network and open the dashboard on the browser.
}
\description{
This function creates an interactive dashboard with two tabs.
The first tab "Not-Isolated Genes" shows a three-colour networks of the not-isolated genes, i.e. the genes sharing
at least one KEGG pathway.
The second tab "Full Network" shows a three-colour networks of all the genes selected by the method (all the not zero signatures).
Different node colors are used to show a probability measure between the genes and the disease under investigation as follows:
Blue color is used for the bottom 501 tissue-specific genes (mapped down genes),
White color is used for genes that are not listed as tissue-specific in the HumanBase database (not mapped genes).
}
