% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapSurvGroup.R
\name{HeatmapSurvGroup}
\alias{HeatmapSurvGroup}
\title{Heatmaps of signature genes in low- and high-risk group}
\usage{
HeatmapSurvGroup(
  x1,
  x2,
  screenVars,
  beta,
  PI.train,
  opt.cutoff,
  PI.test,
  th = 3.5
)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}.}

\item{x2}{input training matrix \eqn{n2xp}.}

\item{screenVars}{screened variables obtained from BMD- or DAD-, or BMD+DAD-screening.}

\item{beta}{regression coefficients.}

\item{PI.train}{prognostic index on training set \eqn{T}.}

\item{PI.test}{prognostic index on testing set \eqn{D}.}

\item{th}{threshold for the \eqn{z}-score. Default is \code{th=3.5}.}
}
\value{
heatmap on training set {T} and testing set {D} divided for high-and low-groups.
}
\description{
This function compute the heatmap of genes selected by the combination of BMD-screening and network methods by dividing the patiens in high- and low-risk survival group.
}
