% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG repository}
\alias{KEGG repository}
\title{KEGG repository}
\format{
Objects of class \code{"data.frame"}.
}
\source{
https://www.genome.jp/kegg/pathway.html
}
\usage{
data(KEGGrepository)
}
\description{
A data frame contains pathway maps representing the a-priori biological knowledge including molecular interaction, reaction and networks relations (KEGG, PPI, etc).
}
\examples{
data(KEGGrepository)
}
\keyword{datasets}
