% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalCoxRanking.R
\name{MarginalCoxRanking}
\alias{MarginalCoxRanking}
\title{Marginal Cox ranking function}
\usage{
MarginalCoxRanking(x, y)
}
\arguments{
\item{x}{input matrix \code{nxp}.}

\item{y}{response variable, \code{y} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}
}
\value{
A data frame composed by the original indices, gene symbols, the marginal regression coefficients, the regression coefficients and the relative p.values.
}
\description{
This function ranks the genes that are highly-correlated with the patient survival according to their corresponding marginal cox utilities. The genes are ordered from the largest to the smallest.
}
\references{
Fan, J., Feng, Y., & Wu, Y. (2010).
High-dimensional variable selection for Cox’s proportional hazards model.\cr
\emph{IMS Collections}, \bold{6}, 70-86.\cr \cr
Saldana, D. F. and Feng, Y. (2018).
SIS: An R package for Sure Independence Screening in Ultrahigh Dimensional Statistical Models.\cr
\emph{Journal of Statistical Software}, \bold{83}, 2, 1-25.\cr\cr
}
