% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkMatrix.R
\name{NetworkMatrix}
\alias{NetworkMatrix}
\title{Functional linkage matrix and Co-expression network matrix}
\usage{
NetworkMatrix(x, th)
}
\arguments{
\item{x}{input matrix \eqn{nxp}}

\item{th}{user-selected threshold}
}
\value{
The following objects are returned as matrices \eqn{pxp}:
\item{FL}{functional linkage matrix (probability matrix)}
\item{adjm}{co-expression network matrix (0/1 matrix)}
}
\description{
This function calculate the functional linkage matrix (\code{FL}) and the co-expression network matrix (\code{adjm}).
Each element of the \code{FL} matrix represents the probability that two adjacent genes are correlated between them.
The \code{adjm} matrix is a binary matrix 0/1 obtained from the \code{FL} matrix depending upon a user-selected threshold.
The \code{adjm} matrix has zero on the diagonal and it is used used for \code{penalty = Net} to calculate Laplacian matrix.
}
\references{
Zhang, W., Ota, T., Shridhar, V., Chien, J., Wu, B., and Kuang, R. (2013),
Network-based survival analysis reveals subnetwork signatures for predicting outcomes of ovarian cancer treatment. \cr
\emph{PLoS computational biology, 9(3), e1002975}\cr \cr
}
