% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NomalizeBetweenData.R
\name{NormalizeBetweenData}
\alias{NormalizeBetweenData}
\title{Normalization between training and testing sets}
\usage{
NormalizeBetweenData(x1, x2, norm.method = "quantile", plot = FALSE)
}
\arguments{
\item{x1}{a matrix of intensities, rows are genes and columns are samples \eqn{pxn1}.}

\item{x2}{a matrix of intensities, rows are genes and columns are samples \eqn{pxn2}.}

\item{plot}{logical flag for plot the normalized data. Default is \code{plot=FALSE}.}

\item{method}{character string specifying the \code{quantile} normalization method.}
}
\value{
The following objects are returned:
\item{x1.norm}{a normalized matrix \eqn{n1xp}}
\item{x2.norm}{a normalized matrix \eqn{n2xp}}
}
\description{
This function performs the normalization between training and testing set. We normalize the data so that the two sets have the same shape (or distribution).
}
\details{
This function uses a quantile normalization in order to make the distributions of training and testing set the same across samples.
The normalization approach used in our package consist of adding by column each sample of the test set to the train set and normalize the new dataset. Then, we take the test column normalized and build the normalized testing set column by column.
This improves the performance and stability of the models and make the two datasets comparable between them.
}
