% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Repository tissue-specific genes}
\alias{Repository tissue-specific genes}
\title{Repository tissue-specific genes}
\format{
Objects of class \code{"data.frame"}.
}
\source{
https://hb.flatironinstitute.org
}
\usage{
data(repositoryDisease)
}
\description{
A repository of prediction scores or posterior probabilities associated to each gene. The higher the probability the stronger the functional relation between the gene and the tissue cancer.
}
\examples{
data(repositoryDisease)
}
\keyword{datasets}
