% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreeningMethod.R
\name{ScreeningMethod}
\alias{ScreeningMethod}
\title{Screening methods (BMD-, DAD- and BMD+DAD-screening)}
\usage{
ScreeningMethod(
  x1,
  myList,
  screening = c("bmd", "dad", "bmd+dad"),
  ranking = NULL,
  thresh = NULL
)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}.}

\item{myList}{list of cancer-specific genes.}

\item{screening}{type of screening (BMD-, DAD- and BMD+DAD-screening).}

\item{ranking}{ranking of genes that are highly-correlated with the patient survival. Default is \code{ranking=NULL}, i.e., for BMD-screening.}

\item{thresh}{number of the ranked genes choosen. Default is \code{thresh=NULL}, i.e., for BMD-screening.}
}
\value{
A object contains the following list of genes:
\item{screenVars}{list of the screened genes.}
}
\description{
This function computes the BMD, DAD and BMD+DAD-screening considering three different ranking cases
in order to select the optimal threshold and optimize data prediction.
}
