% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreeningType.R
\name{ScreeningType}
\alias{ScreeningType}
\title{Screening techniques (DAD- and BMD+DAD-screening)}
\usage{
ScreeningType(
  screenType = c("regCoef", "pValue", "regCoef+pValue"),
  ranking,
  th,
  genesList
)
}
\arguments{
\item{screenType}{three different type of screening based on a, b and c. The alternative are: \code{regCoef}, \code{pValue} or \code{regCoef+pValue}.}

\item{ranking}{the ranking of genes that are highly-correlated with the patient survival.}

\item{th}{the thresholds-dependent genes selected by the users.}

\item{genesList}{the list of genes identified using the BMD-screening.}
}
\value{
A object contains the following two list of genes that are highly-correlated with the patient survival according to the type of ranking selected (a,b or c):
\item{screenVars.DAD}{list of the screened variables using DAD-screening.}
\item{screenVars.BMD.DAD}{list of the screened variables using BMD+DAD-screening.}
}
\description{
This function computes the and BMD+DAD-screening considering three different ranking cases
in order to select the optimal threshold \code{d} and optimize data prediction. The three cases are:
(a) ranking by \verb{regression coefficients}; (b) ranking by \code{p.values}; (c) ranking by \verb{regression coefficients} and \code{p.values}.
}
