% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectOptimalCutoff.R
\name{SelectOptimalCutoff}
\alias{SelectOptimalCutoff}
\title{Select adaptively the PI-optimal cutoff on the training set}
\usage{
SelectOptimalCutoff(x, y, beta)
}
\arguments{
\item{x}{input training matrix \eqn{nxp}.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{beta}{regression coefficients estimated on the training set \eqn{T}.}
}
\value{
The following objects are returned:
\item{df}{data frame abount sample, prognostic index, time, status and group risk for each quantile \eqn{q_{gamma}}, with \eqn{\gamma=0.20, ..., 0.80}.}
\item{summary}{summary information about number of patients at risk, cutoff and p.value for each quantile.}
\item{cutoff.opt}{optimal cutoff selected adaptively.}
\item{p.value}{\eqn{p}-value resulting from the log-rank test (the significance level is \eqn{p}-value < 0.05).}
\item{plots}{survival curves and distribution plot of prognostic index \eqn{PI^{T}}.}
}
\description{
This function computes the optimal cutoff \eqn{PI^{*,T}} on the training set \eqn{T}, i.e., the value that corresponds to the best separation in high-and-low risk group with respect to the log-rank test using the prognostic index \eqn{PI^{T}}.
}
