% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitData.R
\name{SplitData}
\alias{SplitData}
\title{Split dataset in training and testing set}
\usage{
SplitData(x, q)
}
\arguments{
\item{x}{input matrix \eqn{nxp}.}

\item{q}{percentage of the sample size.}
}
\value{
Training and testing set.
}
\description{
This function splits data in training and testing set.
}
