% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tissue Specific functional interaction}
\alias{Tissue Specific functional interaction}
\title{Tissue Specific functional interaction}
\format{
Objects of class \code{"data.frame"}.
}
\source{
https://hb.flatironinstitute.org/download
}
\usage{
data(tissueSpecificEdge)
}
\description{
Top Edges: the network is filtered to only include edges with evidence supporting a tissue-specific functional interaction
[entrez gene id 1][entrez gene id 2][posterior prob.]
}
\examples{
data(tissueSpecificEdge)
}
\keyword{datasets}
