% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationTest.R
\name{ValidationTest}
\alias{ValidationTest}
\title{Prognostic index (PI) and Kaplan-Meier curves on testing set}
\usage{
ValidationTest(x, y, beta, opt.cutoff)
}
\arguments{
\item{x}{input testing matrix \code{nxp}.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{beta}{regression cofficients estimated on the training set \code{T}.}

\item{opt.cutoff}{optimal cutoff selected adaptively on the training set \code{T}.}
}
\value{
The following objects are returned:
\item{df}{data frame about sample, prognostic index, time, status and group risk on testing set \code{D}.}
\item{p.value}{\code{p}-value resulting from the log-rank test (the significance level is \code{p}-value < 0.05).}
\item{plots}{survival curves and distribution plot of prognostic index \code{PI^{D}}.}
}
\description{
This function compute the prognostic index \code{PI^{D}} on the testing set \code{D} using the regression coefficients and the optimal cutoff computed on the training set \code{T}. Then, the kaplan-Meier curves are shown as resulting of the log-rank test between high- and low-risk group.
}
