% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableScreening.R
\name{VariableScreening}
\alias{VariableScreening}
\title{Fit network-regularized Cox regression models on screened data}
\usage{
VariableScreening(x1, y1, x2, y2, screenVars, Omega, alpha, ncv, name)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}.}

\item{y1}{response variable, \code{y1} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{x2}{input testing matrix \eqn{n2xp}.}

\item{y2}{response variable, \code{y2} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{screenVars}{screened variables obtained from DAD-and BMD+DAD-screening by using \code{ScreeningType()} function.}

\item{Omega}{adjacency matrix with zero diagonal and non-negative off-diagonal used to calculate Laplacian matrix.}

\item{alpha}{ratio between \code{L_1} and Laplacian for \code{Net}. Default is \code{alpha = 0.5}.}

\item{ncv}{number of cross-validation performed for tuning optimal parameters (number of folds by default is 5).}

\item{name}{type of screening used from \code{ScreeningType()} function.}
}
\value{
The following objects are returned:
\item{summary}{a data frame composed by number of screened variables (\code{nGenes}), optimal tuning parameters (\code{alpha} and \code{lambda}), number of selected genes (\code{nonZeroBeta}) and the \eqn{p}-value on the testing set (\code{pValue}).}
\item{beta}{a data frame of gene symbols and relative regression coefficients which is also saved in a file \code{txt}.}
\item{PI.train and PI.test}{prognostic index \eqn{PI} computed on the training \eqn{T} and testing set \eqn{D}.}
\item{survival curves}{plot of survival curves.}
\item{summary plots}{two scatter plots: \verb{nGenes vs pValue} and \verb{nGenes vs nonZeroBeta} only if more threshods (\code{nGenes}) are selected}
}
\description{
This function implemets network-regularized Cox regression models which are iteratively trained and validated on different screened sets.
}
