% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VennDiagramScreening.R
\name{VennDiagramScreening}
\alias{VennDiagramScreening}
\title{Venn Diagram screening-network analysis}
\usage{
VennDiagramScreening(genes.bmd, genes.dad, genes.bmd.dad, output = TRUE)
}
\arguments{
\item{genes.bmd}{BMD-screened variables.}

\item{genes.dad}{DAD-screened variables.}

\item{genes.bmd.dad}{BMD+DAD-screened variables.}

\item{output}{logical; if TRUE a PNG file of venn diagram is saved.}
}
\value{
A list of lists which contain the values assigned to each of the areas of a venn diagram is returned.
}
\description{
This function takes the list of high-risk genes selected by each screening-network method and creates a PNG Venn Diagram.
}
