% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosmonetTesting.R
\name{CosmonetTesting}
\alias{CosmonetTesting}
\title{Make predictions on data}
\usage{
CosmonetTesting(x, y, screenVars, beta, opt.cutoff)
}
\arguments{
\item{x}{input testing matrix \code{nxp}. Each row is an observation vector.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{screenVars}{screened variables obtained from BMD- or DAD-, or BMD+DAD-screening.}

\item{beta}{regression coefficients estimated on the training set.}

\item{opt.cutoff}{optimal cutoff selected adaptively on the training set \code{T}.}
}
\value{
The following objects are returned:
\item{df}{data frame composed by prognostic index \code{PI^{D}}, sample, time and status.}
\item{p.value}{resulting from the log-rank test (the significance level is \code{p}-value < 0.05).}
}
\description{
This function performes the model validation and prediction using the testing set \code{D}. It generates the Kaplan-Meier curve resulting from the log-rank test between high and low-risk group and the distribution plot of prognostic index \code{PI^{D}} computed using the signature genes and the optimal cutoff \code{PI^{*,T}} obtained from the training set \code{T}.
}
