% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosmonetTraining.R
\name{CosmonetTraining}
\alias{CosmonetTraining}
\title{Fit network-regularized Cox regression models on training set}
\usage{
CosmonetTraining(
  k,
  x,
  y,
  screenVars,
  family = "Cox",
  penalty = "Net",
  Omega,
  alpha = 0.5,
  lambda = NULL,
  nlambda = 50,
  nfolds = 5,
  foldid = NULL,
  selOptLambda = min("min", "1se"),
  optCutpoint = c("minPValue", "median", "survCutpoint")
)
}
\arguments{
\item{k}{times to loop through cross validation.}

\item{x}{input training matrix \code{nxp}. Each row is an observation vector.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{screenVars}{screened variables obtained from BMD- or DAD-, or BMD+DAD-screening or by the user. A list or a character string can be used.}

\item{family}{Cox proportional hazards regression model. \code{Family=Cox}}

\item{penalty}{penalty type. Can choose \code{Net} where \code{Omega} matrix is requested. For \code{penalty = Net}, the penalty is defined as \eqn{\lambda*{\alpha*||\beta||_1+(1-\lapha)/2*(\beta^{T}L\beta)}},
where \code{L} is a Laplacian matrix calculated from \code{Omega}.}

\item{Omega}{adjacency matrix with zero diagonal and non-negative off-diagonal used to calculate Laplacian matrix.}

\item{alpha}{ratio between \code{L_1} and Laplacian for \code{Net}. Default is \code{alpha = 0.5}.}

\item{lambda}{a user supplied decreasing sequence. If \code{lambda = NULL}, a sequence of lambda is generated based on \code{nlambda} and \code{rlambda} (for more details,  see \code{APML0} package). Supplying a value of lambda overrides this.}

\item{nlambda}{number of lambda values. Default is 50.}

\item{nfolds}{number of folds performed for tuning optimal parameters over runs. Default is \code{nfolds = 5}.}

\item{foldid}{an optional vector of values between 1 and nfolds specifying which fold each observation is in.}

\item{selOptLambda}{a character string for selecting the lambda parameter. Options are \code{min} which uses the regularisation procedure implemented in the APML0 package or \verb{1se"} to select the lambda parameter within one standard error from the optimal value.}

\item{optCutpoint}{a character string for choosing the optimal cutpoint on training set \code{T} based on prognostic index \code{PI^{T}}. Can choose \code{minPValue}, \code{median} and \code{survCutpoint}.}
}
\value{
The following objects are returned:
\item{beta}{a sparse Matrix of coefficients, stored in class\code{dgCMatrix}.}
\item{opt.lambdas}{\eqn{lambda} values based on minimum \code{cvm} over runs and on \code{1se} to select the lambda parameter within one standard error from the optimal value.}
\item{df}{data frame composed by samples, relative prognostic indices, times, status and group risk for each quantile \eqn{q_{gamma}}, with \eqn{\gamma=0.20, ..., 0.80}.}
\item{summary}{summary table on number of patients at risk, cutoff and p.value for each quantile.}
\item{opt.cutoff}{optimal cutoff selected on the training.}
\item{p.value}{resulting from the log-rank test (the significance level is \code{p}-value < 0.05).}
}
\description{
This function fits penalized Cox regression methods in order to incorporate gene regulatory relationships and to select signature genes using the training set \code{T}.
}
