% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateNetwork.R
\name{GenerateNetwork}
\alias{GenerateNetwork}
\title{Generate a three-colour network to investigate genes selected by the network-based Cox model}
\usage{
GenerateNetwork(
  ListOfGenes = listOfGenes,
  ListHeader = TRUE,
  outputFile = "NetworkFile.pdf",
  userWidth = 12,
  userHeight = 6,
  includeLegend = TRUE
)
}
\arguments{
\item{ListOfGenes}{input list of genes.}

\item{ListHeader}{true if the genes' list has a header, false otherwise}

\item{outputFile}{file name for the output network file}

\item{userWidth}{the width of the graphics region in inches.}

\item{userHeight}{the height of the graphics region in inches.}

\item{includeLegend}{if TRUE the legend will be included in the same pdf file, otherwise a new pdf file with the legend will be generated
to run this file: GenerateNetwork()}
}
\description{
This function creates a three-colour network using a list of genes as input (.txt file).
Different node colors are used to show a correlation measure between the genes and the disease under investigation as follows:
Orange nodes represent the genes that are reported by the Functional Map of the Human Genome (Huttenhower, et. al 2008) as strongly correlated to the analysed cancer (p-value <= 0.05);
Green nodes represent the genes that are reported by the Functional Map of the Human Genome as weakly correlated to the analysed cancer (p-value > 0.05);
Purple nodes represent the genes that are not explored by the Functional Map of the Human Genome.
}
