% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalCoxRanking.R
\name{MarginalCoxRanking}
\alias{MarginalCoxRanking}
\title{Marginal Cox ranking function}
\usage{
MarginalCoxRanking(x1, y1)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}.}

\item{y1}{response variable, \code{y1} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}
}
\value{
A data frame composed by the ranking of genes that are highly-correlated with the patient survival, the corresponding regression coefficients and p.values.
}
\description{
This function computes the ranking of genes that are highly-correlated with the patient survival according to their corresponding marginal cox utilities from the largest to the smallest.
}
\references{
Fan, J., Feng, Y., & Wu, Y. (2010).
High-dimensional variable selection for Cox’s proportional hazards model.\cr
\emph{In Borrowing strength: Theory powering applications–a Festschrift for Lawrence D.} Brown (pp. 70-86). Institute of Mathematical Statistics.\cr \cr
}
