% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkCox.R
\name{NetworkCox}
\alias{NetworkCox}
\title{Fit Network-regularized Cox regression models}
\usage{
NetworkCox(x1, y1, penalty = "Net", Omega = NULL, alpha = 0.5, ncv = NULL)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}}

\item{y1}{response variable, \code{y1} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{penalty}{penalty type. Can choose \code{Net} where \code{Omega} matrix is requested.
For \code{penalty = Net}, the penalty is defined as
\eqn{\lambda*{\alpha*||\beta||_1+(1-\lapha)/2*(\beta^{T}L\beta)}},
where L is a Laplacian matrix  calculated from \code{Omega}.}

\item{Omega}{adjacency matrix with zero diagonal and non-negative off-diagonal used to calculate Laplacian matrix.}

\item{alpha}{ratio between \code{L_1} and Laplacian for \code{Net}. Default is \code{alpha = 0.5}.}

\item{ncv}{number of cross-validation performed for tuning optimal parameters (number of folds by default is 5).}
}
\value{
The following objects are returned:
\item{beta}{a matrix of coefficients.}
\item{opt.tuningPars}{optimal tuning parameters.}
\item{nonzerobeta}{non-zero regression coefficients.}
}
\description{
This function fits network-regularized Cox regression methods with net (L1 and Laplacian) penalty. See for more details \code{APML0} package.
}
\references{
Li, X., Xie, S., Zeng, D., Wang, Y. (2018).
Efficient l0-norm feature selection based on augmented and penalized minimization. \cr
\emph{Statistics in medicine, 37(3), 473-486.}
Sun, H., Lin, W., Feng, R., and Li, H. (2014).
Network-regularized high-dimensional cox regression for analysis of genomic data.\cr
\emph{Statistica Sinica}.
Boyd, S., Parikh, N., Chu, E., Peleato, B., & Eckstein, J. (2011).
Distributed optimization and statistical learning via the alternating direction method of multipliers. \cr
\emph{Foundations and Trends in Machine Learning, 3(1), 1-122.}
Friedman, J., Hastie, T., Tibshirani, R. (2010). Regularization paths for generalized linear models via coordinate descent. \cr
\emph{Journal of Statistical Software, Vol. 33(1), 1.} \cr \cr
}
