% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectOptimalCutoff.R
\name{SelectOptimalCutoff}
\alias{SelectOptimalCutoff}
\title{Determine the optimal cutpoint and draw survival curves}
\usage{
SelectOptimalCutoff(
  x,
  y,
  beta,
  optCutpoint = c("minPValue", "median", "survCutpoint")
)
}
\arguments{
\item{x}{input training matrix \code{nxp}.}

\item{y}{response variable, \code{y} should be a two-column data frame with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored. The rownames indicate the sample names ordered as the samples in the input testing matrix.}

\item{beta}{regression coefficients estimated on training set \code{T}.}

\item{optCutpoint}{a character string for choosing the optimal cutpoint on training set \code{T} based on prognostic index \code{PI^{T}}. Can choose \code{minPValue}, \code{median} and \code{survCutpoint}.}
}
\value{
The following objects are returned:
\item{df}{data frame about sample, prognostic index, time, status and group risk for each quantile \eqn{q_{gamma}}, with \eqn{\gamma=0.25, ..., 0.80}.}
\item{summary}{summary information about number of patients at risk, cutoff and p.value for each quantile.}
\item{opt.cutoff}{optimal cutoff selected.}
}
\description{
This function computes the optimal cutoff \code{PI^{*,T}} on the training set \code{T}, i.e., the value that corresponds to the best separation in high-and-low risk group with respect to the log-rank test using the prognostic index \code{PI^{T}}.
}
\details{
We use \code{surv_cutpoint()} to determine the optimal cutpoint and  \code{ggsurvplot()} function to plot survival curves from \code{survminer} package.
}
