% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectOptimalCutoff.R
\name{SelectOptimalCutoff}
\alias{SelectOptimalCutoff}
\title{Select optimal cutoff on the training set}
\usage{
SelectOptimalCutoff(x1, y1, beta)
}
\arguments{
\item{x1}{input training matrix \eqn{n1xp}.}

\item{y1}{response variable, \code{y1} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{beta}{regression coefficients.}
}
\value{
The following objects are returned:
\item{PI.train}{prognostic index \eqn{PI} on the training set \eqn{T}. Compute the quantile \eqn{q_{\gamma}} of \eqn{PI} with \eqn{\gamma = \{0.20, 0.25, 0.30,... 0.80\}}. Each patient \eqn{i} in \eqn{T} is assigned to the high-risk (or low-risk) group if its prognostic index \eqn{PI_i} is above (or below) the \eqn{q_{\gamma}}-quantile.}
\item{cutoff.opt}{optimal cutoff selected adaptively on the training set \eqn{T}. It corresponds to the best separation in high-and-low risk group with respect to the log-rank test.}
\item{p.value}{\eqn{p}-value on the training set (the significance level is \eqn{p}-value < 0.05).}
}
\description{
This function computes the optimal cutoff \eqn{PI^{*,T}} on \eqn{T}, i.e., the value that corresponds to the best separation
in high-and-low risk group with respect to the log-rank test using the training set \eqn{T}.
}
