% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosmonetTesting.R
\name{ValidationTest}
\alias{ValidationTest}
\title{Validation test and Kaplan Meier curve}
\usage{
ValidationTest(
  x2,
  y2,
  beta,
  opt.cutoff,
  screening = c("BMD", "DAD", "BMD+DAD")
)
}
\arguments{
\item{x2}{input testing matrix \eqn{n2xp}. Each row represents an observation, while each column a variable.}

\item{y2}{response variable, \code{y2} should be a three-column matrix with columns named sample, time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{beta}{regression coefficients estimated from variable selection methods on the training set.}

\item{opt.cutoff}{optimal cutoff selected adaptively on the training set \eqn{T}.}

\item{screening}{type of screening applied on the training set \eqn{T}. Can choose BMD, DAD or BMD+DAD.}

\item{screenVars}{screened variables obtained from BMD- or DAD-, or BMD+DAD-screening.}
}
\value{
The following objects are returned:
\item{PI.test}{prognostic index \eqn{PI} computed on the testing set \eqn{D}.}
\item{p.value}{\eqn{p}-value on the testing set \eqn{D} (the significance level is \eqn{p}-value < 0.05).}
\item{HL.groups}{high-risk and low-risk prognosis groups.}
\item{survival curves}{plot of survival curves.}
}
\description{
This function performes the model validation and prediction using the testing set \eqn{D}.
}
