% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationTest.R
\name{ValidationTest}
\alias{ValidationTest}
\title{Validation test}
\usage{
ValidationTest(
  x2,
  y2,
  beta,
  opt.cutoff,
  screening = c("BMD", "DAD", "BMD+DAD")
)
}
\arguments{
\item{x2}{input testing matrix \eqn{n2xp}.}

\item{y2}{response variable, \code{y2} should be a two-column matrix with columns named time and status. The latter is a binary variable, with 1 indicating event, and 0 indicating right censored.}

\item{beta}{regression cofficients estimated from variable selection methods on the screened subset (BMD, DAD or BMD+DAD-screening).}

\item{opt.cutoff}{optimal cutoff selected adaptively on the training set \eqn{T}.}

\item{screening}{type of screening applied on the training set \eqn{T}. Can choose BMD, DAD or BMD+DAD.}
}
\value{
The following objects are returned:
\item{PI.test}{prognostic index \eqn{PI} on the testing set \eqn{D}.}
\item{p.value}{\eqn{p}-value on the testing set \eqn{D} (the significance level is \eqn{p}-value < 0.05).}
\item{HL.groups}{high-and low-risk groups.}
\item{SurvPlot}{plot of survival curves.}
}
\description{
This function performs the validation test using the testing set \eqn{D}, the regression coefficients and the optimal cutoff computed on the training set \eqn{T}.
}
